honksquad: refactor: extract WoundDisplaySystem from SharedWoundSystem
Move display concerns (GetWoundDisplayInfo, GetBleedTier,
UpdateBleedSource) into dedicated WoundDisplaySystem. SharedWoundSystem
now focuses on damage-to-wound pipeline and wound state queries.
